Removes anzii package as it is not needed for this project
The inclusion of the package in question came as a result of anzii-cli command line tool that was used to generate this project repository. The command line tool does not have an implementation suitable for a project of React framework's configs, so I have to do some adjustments. My reason for using this cli tool was to automate common tasks done at the start of every nodejs proect. It also happens to be the tool I have made, so there was also an element of joy involved with using it.