Updates carousel styles with styles made off this repo
Last night I resumed work on the assessment, but instead of resuming it directly on this repo, I decided to clone some content of this repo and started a new one where I would not have to make constant commits. I a couple of reasons for doing this, and ofcourse, this is partly due to the fact that I had an option to do that for this specific assessment I guess.

The points below will be outlining why I took this approach:

- I didn't want to croud my repo with excessive commit messages or history (And I know git has a way of better handling this issue, but I felt time was against me.)

- Working on that new repo would allow me to build the remaining assessment functionality without worrying about switching branches.

- I believed that focusing less on branches and possible commits would buy me more time (And ofcourse, I'm only doing this for this assessment.)